Move calculate_polar_data to helpers.py
